requirements.txt
flask==3.1.3
gunicorn==26.0.0
Pillow==12.2.0
Authlib==1.7.2
requests==2.34.2
redis==7.4.0